+++ service/user/server/src/main/java/com/sparta/user/presentation/request/PointHistoryRequest.java
@@ -2,6 +2,7 @@ package com.sparta.user.presentation.request;
 
 import com.sparta.user.domain.model.vo.PointHistoryType;
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,7 +18,7 @@ public class PointHistoryRequest {
     private Long userId;
     private Long orderId;
     @NotNull
-    private Integer point;
+    private BigDecimal point;
     @NotNull
     private PointHistoryType type;
     @NotNull
